Refactor of code after Pull Request
