+++ src/main/java/com/infoshareacademy/mistrzowieVaBank/form/CustomerForm.java
@@ -3,19 +3,26 @@ package com.infoshareacademy.mistrzowieVaBank.form;
 
 import com.infoshareacademy.mistrzowieVaBank.dto.CustomerInfo;
 
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
+
 public class CustomerForm {
-	 
+    @NotBlank
     private String name;
+    @NotBlank
     private String address;
+    @NotBlank
+    @Email
     private String email;
+    @NotBlank
     private String phone;
- 
+
     private boolean valid;
- 
+
     public CustomerForm() {
- 
+
     }
- 
+
     public CustomerForm(CustomerInfo customerInfo) {
         if (customerInfo != null) {
             this.name = customerInfo.getName();
@@ -24,45 +31,45 @@ public class CustomerForm {
             this.phone = customerInfo.getPhone();
         }
     }
- 
+
     public String getName() {
         return name;
     }
- 
+
     public void setName(String name) {
         this.name = name;
     }
- 
+
     public String getEmail() {
         return email;
     }
- 
+
     public void setEmail(String email) {
         this.email = email;
     }
- 
+
     public String getAddress() {
         return address;
     }
- 
+
     public void setAddress(String address) {
         this.address = address;
     }
- 
+
     public String getPhone() {
         return phone;
     }
- 
+
     public void setPhone(String phone) {
         this.phone = phone;
     }
- 
+
     public boolean isValid() {
         return valid;
     }
- 
+
     public void setValid(boolean valid) {
         this.valid = valid;
     }
- 
+
 }
