+++ src/main/resources/templates/shoppingCart.html
@@ -51,7 +51,7 @@
             <tr >
               <td colspan="7" align="right">
               <input class="btn btn-dark" type="submit" value="Update Quantity" />
-              <a th:href="@{/winelist}" style="margin-left: 20px;">Continue Shopping</a>
+              <a th:href="@{/winelist/page/1}" style="margin-left: 20px;">Continue Shopping</a>
               <a th:href="@{/shoppingCartCustomer}">Checkout</a>
                
               </td>
