+++ src/main/resources/templates/shoppingCartConfirmation.html
@@ -23,7 +23,7 @@
             <li>Quantity: <span th:utext="${myCart.quantityTotal}"></span>$</li>
             <li>Total:
                <span class="total"
-                  th:utext="${#numbers.formatDecimal(myCart.amountTotal,3,2,'COMMA')}">  
+                  th:utext="${myCart.amountTotal} + '$'">
                </span>
             </li>
          </ul>
@@ -55,13 +55,13 @@
                <li>Name: <span th:utext="${cartLineInfo.wineInfo.name}"></span></li>
                <li>Price:
                   <span class="price"
-                     th:utext="${#numbers.formatDecimal(cartLineInfo.wineInfo.price,3,2,'COMMA')}">
+                     th:utext="${cartLineInfo.wineInfo.price} + '$'">
                   </span>
                </li>
                <li>Quantity: <span th:utext="${cartLineInfo.quantity}"></span></li>
                <li>Subtotal:
                   <span class="subtotal"
-                     th:utext="${#numbers.formatDecimal(cartLineInfo.amount,3,2,'COMMA')}">
+                     th:utext="${cartLineInfo.amount} + '$'">
                   </span>
                </li>
             </ul>
